HeroList 컴포넌트로 관심사 분리
